Refactor: Main 클래스를 LibraryApplication으로 명확히 하고, 내부 로직을 InitApplication으로 따로 구분한다. Scanner를 공통 부로 구분해 모든 메서드에서 사용할 수 있도록 한다.
