+++ src/main/java/LibraryApplication.java
@@ -1,0 +1,7 @@
+import common.InitApplication;
+
+public class LibraryApplication {
+    public static void main(String[] args) {
+        InitApplication.run();
+    }
+}
