+++ src/main/java/common/Util.java
@@ -1,0 +1,7 @@
+package common;
+
+import java.util.Scanner;
+
+public class Util {
+    public static Scanner sc = new Scanner(System.in);
+}
